+++ lib/player.ts
@@ -11,6 +11,13 @@ import * as MC from "@minecraft/server";
 export { GameMode };
 export { ScreenDisplay, PlayerLeaveAfterEvent } from "@minecraft/server";
 
+export interface IPlayerSession {
+    /** Called when the player is about to leave. This function will be
+     * called in read-only mode.
+     */
+    destroy(): void;
+}
+
 export class Player extends Entity implements IPreferencesContainer {
     /** Package private: user code should not use this. */
     public constructor(rawPlayer: MC.Player) {
@@ -72,9 +79,72 @@ export class Player extends Entity implements IPreferencesContainer {
             Prefs.encode(ty, prefs));
     }
 
+    /** Obtain a player session object. To use this method, you first need
+     * to register your session class using {@link
+     * World.prototype.usePlayerSessions}.
+     */
+    public getSession<T extends IPlayerSession>(): T {
+        return sessionManager.get(this.id);
+    }
+
     public sendMessage(msg: (RawMessage|string)[]|RawMessage|string): void {
         this.rawPlayer.sendMessage(msg);
     }
+}
+
+/** Package private: user code should not use this. */
+export class SessionManager {
+    readonly #sessions: Map<string, IPlayerSession>; // playerId => IPlayerSession
+    #ctor?: new (player: Player) => IPlayerSession;
+
+    private constructor() {
+        this.#sessions = new Map();
+    }
+
+    public set "class"(ctor: new (player: Player) => IPlayerSession) {
+        if (this.#ctor)
+            throw new Error("A session class cannot be changed once it's set");
+        else
+            this.#ctor = ctor;
+    }
+
+    public create(player: Player): IPlayerSession {
+        if (!this.#ctor)
+            throw new Error("No session classes have been set");
+
+        if (this.#sessions.has(player.id))
+            throw new Error(`Duplicate session for player ${player.id} (${player.name})`);
+
+        const session = new (this.#ctor)(player);
+        this.#sessions.set(player.id, session);
+        return session;
+    }
+
+    public destroy(playerId: string) {
+        const session = this.#sessions.get(playerId);
+        if (session) {
+            this.#sessions.delete(playerId);
+            session.destroy();
+        }
+        else {
+            throw new Error(`Session not found for player ${playerId}`);
+        }
+    }
+
+    public "get"(playerId: string): IPlayerSession {
+        const session = this.#sessions.get(playerId);
+        if (session)
+            return session;
+        else
+            throw new Error(`Session not found for player ${playerId}`);
+    }
+}
+/** Package private: user code should not use this. */
+// @ts-expect-error: Intentionally calling a private constructor.
+export const sessionManager = new SessionManager();
+
+export interface PlayerLeaveBeforeEvent {
+    readonly player: Player;
 }
 
 export interface PlayerSpawnAfterEvent {