+++ lib/world/events.ts
@@ -7,7 +7,7 @@ import { Dimension } from "../dimension.js";
 import { Entity, ItemUseAfterEvent, EntityDieAfterEvent, EntityEventOptions,
          entityEventOptionsToRaw } from "../entity.js";
 import { ItemStack } from "../item/stack.js";
-import { Player, PlayerSpawnAfterEvent } from "../player.js"
+import { Player, PlayerLeaveBeforeEvent, PlayerSpawnAfterEvent } from "../player.js"
 import { Wrapper } from "../wrapper.js";
 import * as MC from "@minecraft/server";
 
@@ -82,6 +82,7 @@ export class WorldAfterEvents extends Wrapper<MC.WorldAfterEvents> {
 
 export class WorldBeforeEvents extends Wrapper<MC.WorldBeforeEvents> {
     public readonly playerBreakBlock: IEventSignal<PlayerBreakBlockBeforeEvent>;
+    public readonly playerLeave: IEventSignal<PlayerLeaveBeforeEvent>;
 
     /// Package private
     public constructor(rawEvents: MC.WorldBeforeEvents) {
@@ -89,6 +90,13 @@ export class WorldBeforeEvents extends Wrapper<MC.WorldBeforeEvents> {
         this.playerBreakBlock = new GluedEventSignalWithoutOptions(
             this.raw.playerBreakBlock,
             (rawEv: MC.PlayerBreakBlockBeforeEvent) => new PlayerBreakBlockBeforeEvent(rawEv));
+        this.playerLeave = new GluedEventSignalWithoutOptions(
+            this.raw.playerLeave,
+            (rawEv: MC.PlayerLeaveBeforeEvent) => {
+                return {
+                    player: new Player(rawEv.player)
+                };
+            });
     }
 }